Resolve conflict between BuildSyntax.dependsOn and Scoped.dependsOn
